+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.7.2] Q1 2023
- - fixed issue when exclude_network_filter is empty
+ - report network status:
+   - switched to include filter
+   - fixed issue with empty filter
 
 ## [0.7.1] Q1 2023
  - added audit exception for RUSTSEC-2020-0071